+++ src/components/utils/src/file_system.cc
@@ -54,7 +54,7 @@ uint64_t file_system::GetAvailableDiskSpace(const std::string& path) {
   }
 }
 
-uint32_t file_system::FileSize(const std::string &path) {
+int64_t file_system::FileSize(const std::string &path) {
   if (file_system::FileExists(path)) {
     struct stat file_info;
     memset(reinterpret_cast<void*>(&file_info), 0, sizeof(file_info));
